refactor(ts): extract buildFilePrompt to file_prompt.ts (mirror #146)
